Organiza model de usuários.
thiagobitencourt/webdev-2#55
