+++ MatheusEduardo/routes/usuarios.js
@@ -1,9 +1,9 @@
 const UsuarioModel = require('../models/usuario');
 const ObjectId = require('mongoose').Types.ObjectId;
 
-const UsuarioDAO = require('../usuarioDAO');
-const usuarioRepo = new UsuarioDAO();
-//TODO: Trocar chamadas do usuarioRepo para chamadas do model correspondente.
+const Usuario = require('../models/usuario');
+const usuarioRepo = new Usuario();
+
 module.exports = {
 
   autenticarUsuario: ((req, res) => {
